add Mutation converter tests (5 new files) and fix 12 bugs
New tests: AvroToMutationConverterTest, RowToMutationConverterTest,
JsonToMutationConverterTest, EntityToMutationConverterTest,
StructToRowConverterTest (31 tests; these converters feed the Spanner
sink and were 0-4% covered).

Fixes: time-millis int overflow in AvroToMutationConverter; TIME fields
read as joda Instant instead of LocalTime and null INT16/INT32 NPE in
RowToMutationConverter; null handling, datetime-array wrong element
access, and outer-array logical type lookup in JsonToMutationConverter;
inverted __key__ name/id condition in EntityToMutationConverter (name
keys always wrote key 0); FLOAT32 read as double, nested struct passed
outer struct to recursion (all-null rows), and returned-not-thrown
exception in StructToRowConverter.

All fixes verified fail-first; regression tests added.

+++ src/main/java/com/mercari/solution/util/schema/converter/AvroToMutationConverter.java
@@ -542,7 +542,7 @@ public class AvroToMutationConverter {
                 if(LogicalTypes.date().equals(fieldSchema.getLogicalType())) {
                     yield Value.date(isNull ? null : convertEpochDaysToDate(intValue));
                 } else if(LogicalTypes.timeMillis().equals(fieldSchema.getLogicalType())) {
-                    yield Value.string(isNull ? null : LocalTime.ofNanoOfDay(((intValue) * 1000 * 1000)).format(DateTimeFormatter.ISO_LOCAL_TIME));
+                    yield Value.string(isNull ? null : LocalTime.ofNanoOfDay(intValue.longValue() * 1000L * 1000L).format(DateTimeFormatter.ISO_LOCAL_TIME));
                 } else {
                     yield Value.int64(isNull ? null : ((Integer) object).longValue());
                 }
@@ -626,7 +626,7 @@ public class AvroToMutationConverter {
                             yield Value.stringArray(isNull ? new ArrayList<>() : ((List<Integer>)object)
                                     .stream()
                                     .filter(Objects::nonNull)
-                                    .map(i -> LocalTime.ofNanoOfDay(((i) * 1000 * 1000)).format(DateTimeFormatter.ISO_LOCAL_TIME))
+                                    .map(i -> LocalTime.ofNanoOfDay(i.longValue() * 1000L * 1000L).format(DateTimeFormatter.ISO_LOCAL_TIME))
                                     .collect(Collectors.toList()));
                         } else {
                             yield Value.int64Array(isNull ? new ArrayList<>() : ((List<Integer>)object)
